🚚 Move large functions to his own files
