+++ apps/contracts/src/tests/mainTest.ts
@@ -23,8 +23,9 @@ import {
   withdrawFromVault
 } from "./vault.js";
 import { green, purple, red, usdcAddress, xtarAddress, yellow } from "./common.js";
-import { testVaultOneAssetOneStrategy, twoAssetsOneStrategySuccess } from "./vault/one_strategy.js";
+import { testVaultOneAssetOneStrategy } from "./vault/one_asset_one_strategy.js";
 import { extractAddresses } from "./vault/utils.js";
+import { twoAssetsOneStrategySuccess } from "./vault/two_assets_one_strategy.js";
 
 const args = process.argv.slice(2);
 const network = args[0];
@@ -107,245 +108,6 @@ async function prepareEnvironment() {
   }
 }
 
-async function testVaultOneStrategy() {
-  console.log(yellow, "---------------------------------------");
-  console.log(yellow, "Running one strategy vault tests");
-  console.log(yellow, "---------------------------------------");
-
-  // deploy vault
-  console.log(purple, "---------------------------------------");
-  console.log(purple, "Deploying vault with one strategy");
-  console.log(purple, "---------------------------------------");
-  const { 
-    address: vaultAddress, 
-    instructions:deployInstructions, 
-    readBytes:deployReadBytes, 
-    writeBytes:deployWriteBytes} = await deployVault(
-    addressBook,
-    oneStrategyParams,
-    "TestVault",
-    "TSTV"
-  );
-  console.log(vaultAddress)
-
-  console.log(yellow, "---------------------------------------");
-  console.log(yellow, "Fetching balances");
-  console.log(yellow, "---------------------------------------");
-
-  const idleFundsBeforeDeposit = await fetchParsedCurrentIdleFunds(
-    vaultAddress,
-    extractAddresses(oneStrategyParams),
-    testUser
-  );
-  const investedFundsBeforeDeposit = await fetchCurrentInvestedFunds(
-    vaultAddress,
-    testUser
-  );
-  const hodlBalanceBeforeDeposit = await checkUserBalance(
-    addressBook.getContractId("hodl_strategy"),
-    vaultAddress,
-    testUser
-  );
-
-  // deposit to vault
-
-  const {
-    instructions: depositInstructions,
-    readBytes:depositReadBytes,
-    writeBytes:depositWriteBytes,
-  } = await depositToVault(vaultAddress, [987654321], testUser);
-
-  console.log(yellow, "---------------------------------------");
-  console.log(yellow, "Fetching balances");
-  console.log(yellow, "---------------------------------------");
-
-  const idleFundsAfterDeposit = await fetchParsedCurrentIdleFunds(
-    vaultAddress,
-    extractAddresses(oneStrategyParams),
-    testUser
-  );
-  const investedFundsAfterDeposit = await fetchCurrentInvestedFunds(
-    vaultAddress,
-    testUser
-  );
-  const hodlBalanceAfterDeposit = await checkUserBalance(
-    addressBook.getContractId("hodl_strategy"),
-    vaultAddress,
-    testUser
-  );
-
-  // withdraw from vault
-  const {
-    instructions: withdrawInstructions,
-    readBytes:withdrawReadBytes,
-    writeBytes:withdrawWriteBytes,
-  } = await withdrawFromVault(vaultAddress, 65_0_000, testUser);
-
-  console.log(yellow, "---------------------------------------");
-  console.log(yellow, "Fetching balances");
-  console.log(yellow, "---------------------------------------");
-
-  const idleFundsAfterWithdraw = await fetchParsedCurrentIdleFunds(
-    vaultAddress,
-    extractAddresses(oneStrategyParams),
-    testUser
-  );
-  const investedFundsAfterWithdraw = await fetchCurrentInvestedFunds(
-    vaultAddress,
-    testUser
-  );
-  const hodlBalanceAfterWithdraw = await checkUserBalance(
-    addressBook.getContractId("hodl_strategy"),
-    vaultAddress,
-    testUser
-  );
-
-  // invest in vault
-  console.log(purple, "---------------------------------------");
-  console.log(purple, "Investing in vault");
-  console.log(purple, "---------------------------------------");
-
-  const investArgs: Instruction[] = [
-    {
-      type: "Invest",
-      strategy: addressBook.getContractId("hodl_strategy"),
-      amount: BigInt(43_0_0),
-    },
-  ];
-
-  const { 
-    result: investResult,
-    instructions: investInstructions,
-    readBytes:investReadBytes,
-    writeBytes:investWriteBytes
-  } = await rebalanceVault(
-    vaultAddress,
-    investArgs,
-    manager
-  );
-  console.log(yellow, "---------------------------------------");
-  console.log(yellow, "Fetching balances");
-  console.log(yellow, "---------------------------------------");
-  const idleFundsAfterInvest = await fetchParsedCurrentIdleFunds(
-    vaultAddress,
-    extractAddresses(oneStrategyParams),
-    testUser
-  );
-  const investedFundsAfterInvest = await fetchCurrentInvestedFunds(
-    vaultAddress,
-    testUser
-  );
-  const hodlBalanceAfterInvest = await checkUserBalance(
-    addressBook.getContractId("hodl_strategy"),
-    vaultAddress,
-    testUser
-  );
-
-  // rebalance vault
-
-  console.log(purple, "---------------------------------------");
-  console.log(purple, "Rebalancing vault"); 
-  console.log(purple, "---------------------------------------");
-
-  const rebalanceArgs: Instruction[] = [
-    {
-      type: "Invest",
-      strategy: addressBook.getContractId("hodl_strategy"),
-      amount: BigInt(7_0_000),
-    },
-    {
-      type: "Unwind",
-      strategy: addressBook.getContractId("hodl_strategy"),
-      amount: BigInt(6_0_00),
-    },
-  ];
-
-  const mappedParams = mapInstructionsToParams(rebalanceArgs);
-
-  const { 
-    result: rebalanceResult,
-    instructions: rebalanceInstructions,
-    readBytes:rebalanceReadBytes,
-    writeBytes:rebalanceWriteBytes
-  } = await rebalanceVault(
-    vaultAddress,
-    investArgs,
-    manager
-  );
-
-  console.log(yellow, "---------------------------------------");
-  console.log(yellow, "Fetching balances");
-  console.log(yellow, "---------------------------------------");
-  const idleFundsAfterRebalance = await fetchParsedCurrentIdleFunds(
-    vaultAddress,
-    extractAddresses(oneStrategyParams),
-    testUser
-  );
-  const investedFundsAfterRebalance = await fetchCurrentInvestedFunds(
-    vaultAddress,
-    testUser
-  );
-  const hodlBalanceAfterRebalance = await checkUserBalance(
-    addressBook.getContractId("hodl_strategy"),
-    vaultAddress,
-    testUser
-  );
-
-  const tableData = {
-    hodlStrategy: {
-      "Balance before deposit": hodlBalanceBeforeDeposit,
-      "Balance after deposit": hodlBalanceAfterDeposit,
-      "Balance after withdraw": hodlBalanceAfterWithdraw,
-      "Balance after invest": hodlBalanceAfterInvest,
-      "Balance after rebalance": hodlBalanceAfterRebalance,
-    },
-    "Invested funds": {
-      "Balance before deposit": investedFundsBeforeDeposit[0].amount,
-      "Balance after deposit": investedFundsAfterDeposit[0].amount,
-      "Balance after withdraw": investedFundsAfterWithdraw[0].amount,
-      "Balance after invest": investedFundsAfterInvest[0].amount,
-      "Balance after rebalance": investedFundsAfterRebalance[0].amount,
-    },
-    "Idle funds": {
-      "Balance before deposit": idleFundsBeforeDeposit[0].amount,
-      "Balance after deposit": idleFundsAfterDeposit[0].amount,
-      "Balance after withdraw": idleFundsAfterWithdraw[0].amount,
-      "Balance after invest": idleFundsAfterInvest[0].amount,
-      "Balance after rebalance": idleFundsAfterRebalance[0].amount,
-    },
-  };
-  console.table(tableData);
-
-  const budgetData = {
-    deploy: {
-      instructions: deployInstructions,
-      readBytes: deployReadBytes,
-      writeBytes: deployWriteBytes,
-    },
-    deposit: {
-      instructions: depositInstructions,
-      readBytes: depositReadBytes,
-      writeBytes: depositWriteBytes,
-    },
-    withdraw: {
-      instructions: withdrawInstructions,
-      readBytes: withdrawReadBytes,
-      writeBytes: withdrawWriteBytes,
-    },
-    invest: {
-      instructions: investInstructions,
-      readBytes: investReadBytes,
-      writeBytes: investWriteBytes,
-    },
-    rebalance: {
-      instructions: rebalanceInstructions,
-      readBytes: rebalanceReadBytes,
-      writeBytes: rebalanceWriteBytes,
-    },
-  }
-  console.table(budgetData);
-  return {tableData, budgetData};
-}
 
 async function testVaultTwoStrategies() {
   console.log(yellow, "---------------------------------------");
@@ -629,286 +391,6 @@ async function testVaultTwoStrategies() {
     },
   }
 
-  console.table(tableData);
-  console.table(budgetData);
-  return {tableData, budgetData};
-}
-
-async function testVaultTwoAssetsOneStrategy() {
-  console.log(yellow, "---------------------------------------");
-  console.log(yellow, "Running two strategies vault tests");
-  console.log(yellow, "---------------------------------------");
-  // deploy vault
-
-  console.log(purple, "---------------------------------------");
-  console.log(purple, "Deploying vault with two strategies");
-  console.log(purple, "---------------------------------------");
-  const {
-    address:vaultAddress,
-    instructions:deployInstructions,
-    readBytes:deployReadBytes,
-    writeBytes:deployWriteBytes
-  } = await deployVault(
-    addressBook,
-    twoAssetOneStrategyParams,
-    "TestVault",
-    "TSTV"
-  );
-
-  console.log(yellow, "---------------------------------------");
-  console.log(yellow, "Fetching balances");
-  console.log(yellow, "---------------------------------------");
-
-  const idleFundsBeforeDeposit = await fetchParsedCurrentIdleFunds(
-    vaultAddress,
-    extractAddresses(twoAssetOneStrategyParams),
-    testUser
-  );
-  const investedFundsBeforeDeposit = await fetchCurrentInvestedFunds(
-    vaultAddress,
-    testUser
-  );
-  const fixedUsdcBalanceBeforeDeposit = await checkUserBalance(
-    addressBook.getContractId("fixed_usdc_strategy"),
-    vaultAddress,
-    testUser
-  );
-  const fixedXtarBalanceBeforeDeposit = await checkUserBalance(
-    addressBook.getContractId("fixed_xtar_strategy"),
-    vaultAddress,
-    testUser
-  );
-
-  // deposit to vault
-
-  console.log(purple, "---------------------------------------");
-  console.log(purple, "Depositing to vault");
-  console.log(purple, "---------------------------------------");
-
-  const {
-    instructions:depositInstructions,
-    readBytes:depositReadBytes,
-    writeBytes:depositWriteBytes
-  } = await depositToVault(vaultAddress, [98_7_654_321, 98_7_654_321], testUser);
-
-  console.log(yellow, "---------------------------------------");
-  console.log(yellow, "Fetching balances");
-  console.log(yellow, "---------------------------------------");
-
-  const idleFundsAfterDeposit = await fetchParsedCurrentIdleFunds(
-    vaultAddress,
-    extractAddresses(twoAssetOneStrategyParams),
-    testUser
-  );
-  const investedFundsAfterDeposit = await fetchCurrentInvestedFunds(
-    vaultAddress,
-    testUser
-  );
-  const fixedUsdcBalanceAfterDeposit = await checkUserBalance(
-    addressBook.getContractId("fixed_usdc_strategy"),
-    vaultAddress,
-    testUser
-  );
-  const fixedXtarBalanceAfterDeposit = await checkUserBalance(
-    addressBook.getContractId("fixed_xtar_strategy"),
-    vaultAddress,
-    testUser
-  );
-
-  // withdraw from vault
-
-  console.log(purple, "---------------------------------------");
-  console.log(purple, "Withdrawing from vault");
-  console.log(purple, "---------------------------------------");
-
-  const {
-    instructions:withdrawInstructions,
-    readBytes:withdrawReadBytes,
-    writeBytes:withdrawWriteBytes
-  } = await withdrawFromVault(vaultAddress, 7_0_000_000, testUser);
-
-  const idleFundsAfterWithdraw = await fetchParsedCurrentIdleFunds(
-    vaultAddress,
-    extractAddresses(twoAssetOneStrategyParams),
-    testUser
-  );
-  const investedFundsAfterWithdraw = await fetchCurrentInvestedFunds(
-    vaultAddress,
-    testUser
-  );
-  const fixedUsdcBalanceAfterWithdraw = await checkUserBalance(
-    addressBook.getContractId("fixed_usdc_strategy"),
-    vaultAddress,
-    testUser
-  );
-  const fixedXtarBalanceAfterWithdraw = await checkUserBalance(
-    addressBook.getContractId("fixed_xtar_strategy"),
-    vaultAddress,
-    testUser
-  );
-
-  // invest in vault
-  console.log(purple, "---------------------------------------");
-  console.log(purple, "Investing in vault");
-  console.log(purple, "---------------------------------------");
-
-  const investArgs: Instruction[] = [
-    {
-      type: "Invest",
-      strategy: addressBook.getContractId("fixed_xtar_strategy"),
-      amount: BigInt(10_000_000),
-    },
-    {
-      type: "Invest",
-      strategy: addressBook.getContractId("fixed_usdc_strategy"),
-      amount: BigInt(10_000_000),
-    },
-  ];
-
-  const {
-    instructions:investInstructions,
-    readBytes:investReadBytes,
-    writeBytes:investWriteBytes
-  } = await rebalanceVault(vaultAddress, investArgs, manager);
-
-  console.log(yellow, "---------------------------------------");
-  console.log(yellow, "Fetching balances");
-  console.log(yellow, "---------------------------------------");
-
-  const idleFundsAfterInvest = await fetchParsedCurrentIdleFunds(
-    vaultAddress,
-    extractAddresses(twoAssetOneStrategyParams),
-    testUser
-  );
-  const investedFundsAfterInvest = await fetchCurrentInvestedFunds(
-    vaultAddress,
-    testUser
-  );
-  const fixedUsdcBalanceAfterInvest = await checkUserBalance(
-    addressBook.getContractId("fixed_usdc_strategy"),
-    vaultAddress,
-    testUser
-  );
-  const fixedXtarBalanceAfterInvest = await checkUserBalance(
-    addressBook.getContractId("fixed_xtar_strategy"),
-    vaultAddress,
-    testUser
-  );
-
-  // rebalance vault
-  console.log(purple, "---------------------------------------");
-  console.log(purple, "Rebalancing vault");
-  console.log(purple, "---------------------------------------");
-
-  const rebalanceArgs: Instruction[] = [
-    {
-        type: "SwapExactIn",
-        token_in: xtarAddress.toString(),
-        token_out: usdcAddress.toString(),
-        amount_in: BigInt(2_0_000_000),
-        amount_out_min: BigInt(1_0_000_000),
-        deadline: BigInt(Math.floor(Date.now() / 1000) + 3600),
-    },
-    {
-      type: "SwapExactOut",
-      token_in: usdcAddress.toString(),
-      token_out: xtarAddress.toString(),
-      amount_out: BigInt(1_000_000),
-      amount_in_max: BigInt(1_000_000),
-      deadline: BigInt(Math.floor(Date.now() / 1000) + 3600),
-    }
-  ];
-
-  const {
-      instructions:rebalanceInstructions,
-      readBytes:rebalanceReadBytes,
-      writeBytes:rebalanceWriteBytes
-  } = await rebalanceVault(vaultAddress, rebalanceArgs, manager);
-    
-
-  console.log(yellow, "---------------------------------------");
-  console.log(yellow, "Fetching balances");
-  console.log(yellow, "---------------------------------------");
-  const idleFundsAfterRebalance = await fetchParsedCurrentIdleFunds(
-    vaultAddress,
-    extractAddresses(twoAssetOneStrategyParams),
-    testUser
-  );
-  const investedFundsAfterRebalance = await fetchCurrentInvestedFunds(
-    vaultAddress,
-    testUser
-  );
-  const fixedUsdcBalanceAfterRebalance = await checkUserBalance(
-    addressBook.getContractId("fixed_usdc_strategy"),
-    vaultAddress,
-    testUser
-  );
-  const fixedXtarBalanceAfterRebalance = await checkUserBalance(
-    addressBook.getContractId("fixed_xtar_strategy"),
-    vaultAddress,
-    testUser
-  );
-
-  console.log(green, "---------------------------------------");
-  console.log(green, "Tests completed successfully");
-  console.log(green, "---------------------------------------");
-  const tableData = {
-    fixedUsdcStrategy: {
-      "Balance before deposit": fixedUsdcBalanceBeforeDeposit,
-      "Balance after deposit": fixedUsdcBalanceAfterDeposit,
-      "Balance after withdraw": fixedUsdcBalanceAfterWithdraw,
-      "Balance after invest": fixedUsdcBalanceAfterInvest,
-      "Balance after rebalance": fixedUsdcBalanceAfterRebalance,
-    },
-    fixedXtarStrategy: {
-      "Balance before deposit": fixedXtarBalanceBeforeDeposit,
-      "Balance after deposit": fixedXtarBalanceAfterDeposit,
-      "Balance after withdraw": fixedXtarBalanceAfterWithdraw,
-      "Balance after invest": fixedXtarBalanceAfterInvest,
-      "Balance after rebalance": fixedXtarBalanceAfterRebalance,
-    },
-    "Invested funds": {
-      "Balance before deposit": investedFundsBeforeDeposit[0].amount,
-      "Balance after deposit": investedFundsAfterDeposit[0].amount,
-      "Balance after withdraw": investedFundsAfterWithdraw[0].amount,
-      "Balance after invest": investedFundsAfterInvest[0].amount,
-      "Balance after rebalance": investedFundsAfterRebalance[0].amount,
-    },
-    "Idle funds": {
-      "Balance before deposit": idleFundsBeforeDeposit[0].amount,
-      "Balance after deposit": idleFundsAfterDeposit[0].amount,
-      "Balance after withdraw": idleFundsAfterWithdraw[0].amount,
-      "Balance after invest": idleFundsAfterInvest[0].amount,
-      "Balance after rebalance": idleFundsAfterRebalance[0].amount,
-    },
-  };
-  const budgetData = {
-    deploy: {
-      instructions: deployInstructions,
-      readBytes: deployReadBytes,
-      writeBytes: deployWriteBytes,
-    },
-    deposit: {
-      instructions: depositInstructions,
-      readBytes: depositReadBytes,
-      writeBytes: depositWriteBytes,
-    },
-    withdraw: {
-      instructions: withdrawInstructions,
-      readBytes: withdrawReadBytes,
-      writeBytes: withdrawWriteBytes,
-    },
-    invest: {
-      instructions: investInstructions,
-      readBytes: investReadBytes,
-      writeBytes: investWriteBytes,
-    },
-    rebalance: {
-      instructions: rebalanceInstructions,
-      readBytes: rebalanceReadBytes,
-      writeBytes: rebalanceWriteBytes,
-    },
-  }
   console.table(tableData);
   console.table(budgetData);
   return {tableData, budgetData};
